packages/anumate-tracing/pyproject.toml
[tool.poetry]
name = "anumate-tracing"
version = "0.1.0"
description = "OpenTelemetry tracing utilities for the Anumate platform."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-api = "^1.12.0"
opentelemetry-sdk = "^1.12.0"
opentelemetry-instrumentation-fastapi = "^0.33b0"
opentelemetry-instrumentation-requests = "^0.33b0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
